rename strategy for mancala capture
