+++ mancalacapture/mancala_cap_strategy.py
